move Debye_length tests into test_lengths.py
